lots of changes

+++ app/controllers/users.controller.ts
@@ -1,5 +1,4 @@
-import { getMethodActName, mapRequestToQueryParams } from '#lib/utils_server'
-import PermissionCheckService from '#services/permission_check.service'
+import { getMethodActName, mapRequestToQueryParams, throwForbidden } from '#lib/utils'
 import UserService from '#services/user.service'
 import { createEditUserValidator } from '#validators/users'
 
@@ -8,23 +7,16 @@ import type { HttpContext } from '@adonisjs/core/http'
 
 @inject()
 export default class UsersController {
-  constructor(
-    protected service: UserService,
-    protected permChecker: PermissionCheckService
-  ) {}
+  constructor(protected service: UserService) {}
 
-  async viewList({ request, response, auth }: HttpContext) {
+  async viewList({ request, response, bouncer, inertia }: HttpContext) {
+    if (await bouncer.with('UserPolicy').denies('view')) return throwForbidden()
+
     try {
-      await this.permChecker.checkPerm(auth.user!, 'user.view')
-
       const q = mapRequestToQueryParams(request)
       const data = await this.service.index(q)
 
-      return response.status(200).json({
-        status: 'success',
-        message: 'Successfully fetched users.',
-        data: data,
-      })
+      return inertia.render('dashboard/users', data)
     } catch (error) {
       return response.status(error.status || 500).json({
         status: 'error',
@@ -35,13 +27,15 @@ export default class UsersController {
 
   // if POST request -> create
   // if PATCH request -> update
-  async storeOrUpdate({ request, response, auth }: HttpContext) {
+  async storeOrUpdate({ bouncer, request, response }: HttpContext) {
     try {
-      await this.permChecker.checkPermInMethod(auth.user!, 'user.create', request, 'POST')
-      await this.permChecker.checkPermInMethod(auth.user!, 'user.update', request, 'PATCH')
+      const payload = await request.validateUsing(createEditUserValidator)
+      if (await bouncer.with('UserPolicy').denies('create', payload, request))
+        return throwForbidden()
+      if (await bouncer.with('UserPolicy').denies('update', payload, request))
+        return throwForbidden()
 
-      const data = await request.validateUsing(createEditUserValidator)
-      await this.service.createEdit(data)
+      await this.service.createUpdate(payload)
 
       return response.status(200).json({
         status: 'success',
@@ -55,11 +49,11 @@ export default class UsersController {
     }
   }
 
-  async destroy({ response, params, auth }: HttpContext) {
+  async destroy({ response, params, bouncer }: HttpContext) {
     try {
-      await this.permChecker.checkPerm(auth.user!, 'user.delete')
-      const id = params.id
+      if (await bouncer.with('UserPolicy').denies('delete')) return throwForbidden()
 
+      const id = params.id
       await this.service.deleteUser(id)
 
       return response.status(200).json({
